🎨fix(domain, constants):금액을 double에서 int로 통일
모든 금액이 100원 단위이므로 double에서 int로 통일시킴.

+++ src/main/java/christmas/constants/Amount.java
@@ -13,9 +13,9 @@ public enum Amount {
     MINIMUM_SANTA_AMOUNT(20000); // 산타 뱃지 금액 - 20,000원
     
 
-    public final double amount;
+    public final int amount;
 
-    Amount(double amount) {
+    Amount(int amount) {
         this.amount = amount;
     }
 }
