+++ src/main/java/christmas/domain/ChristmasDdayEvent.java
@@ -11,12 +11,12 @@ public class ChristmasDdayEvent extends Discount implements Event {
     }
 
     @Override
-    public boolean isValidEventAmount(double totalAmount) {
+    public boolean isValidEventAmount(int totalAmount) {
         return totalAmount >= Amount.MINIMUM_AMOUNT_FOR_EVENT.amount;
     }
 
     @Override
-    public double calculateDiscountAmount(Order order) {
+    public int calculateDiscountAmount(Order order) {
         return Amount.INITIAL_CHRISTMAS_DDAY_DISCOUNT_AMOUNT.amount + Amount.CHRISTMAS_DDAY_DISCOUNT_INCREMENT_AMOUNT_BY_DAY.amount * (order.getReservationDate() - Day.CHRISTMAS_D_DAY_EVENT_START_DAY.getDay());
     }
