+++ src/main/java/christmas/domain/SpecialEvent.java
@@ -13,11 +13,11 @@ public class SpecialEvent extends Discount implements Event {
     }
 
     @Override
-    public boolean isValidEventAmount(double totalAmount) {
+    public boolean isValidEventAmount(int totalAmount) {
         return totalAmount >= Amount.MINIMUM_AMOUNT_FOR_EVENT.amount;
     }
 
-    public double calculateDiscountAmount(Order order) {
+    public int calculateDiscountAmount(Order order) {
         return Amount.SPECIAL_DISCOUNT_AMOUNT.amount;
     }
